Adding fully-functioning withdraw code and tests.
Usually I like to do more TDD but I'm not quite experienced enough with the Solana and Anchor frameworks to
write a full test suite off the bat. So I ended up doing a lot of the program code in tandem with the tests
as I figured out the requisite API calls.